Conflicts:
	app/models/services/ElasticsearchService.java
	app/views/searchresult.scala.html

+++ app/controllers/GroupController.java
@@ -17,6 +17,7 @@ import play.mvc.Call;
 import play.mvc.Result;
 import play.mvc.Security;
 import views.html.Group.*;
+import views.html.Group.snippets.streamRaw;
 
 
 @Transactional
@@ -40,12 +41,13 @@ public class GroupController extends BaseController {
     @Transactional(readOnly=true)
     public static Result view(Long id) {
         Group group = Group.findById(id);
+
         if (group == null) {
-            Logger.error("No group found with id: " + id);
+            flash("error", Messages.get("group.group_not_found"));
             return redirect(controllers.routes.GroupController.index());
         }
         if (Secured.viewGroup(group)) {
-            return redirect(routes.GroupController.stream(group.id, PAGE));
+            return redirect(routes.GroupController.stream(group.id, PAGE, false));
         }
         Navigation.set(Level.GROUPS, "Info", group.title, controllers.routes.GroupController.view(group.id));
 
@@ -53,36 +55,44 @@ public class GroupController extends BaseController {
     }
 
 	@Transactional(readOnly=true)
-	public static Result stream(Long id, int page) {
+	public static Result stream(Long id, int page, boolean raw) {
 		Group group = Group.findById(id);
 
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
         if(!Secured.viewGroup(group)){
 			return redirect(routes.GroupController.view(group.id));
 		}
 
-        Navigation.set(Level.GROUPS, "Newsstream", group.title, controllers.routes.GroupController.stream(group.id, PAGE));
+        Navigation.set(Level.GROUPS, "Newsstream", group.title, controllers.routes.GroupController.stream(group.id, PAGE, false));
         List<Post> posts = Post.getPostsForGroup(group, LIMIT, page);
 
-        return ok(stream.render(group, posts, postForm, Post.countPostsForGroup(group), LIMIT, page));
-
+        if(raw) {
+            return ok(streamRaw.render(group, posts, postForm, Post.countPostsForGroup(group), LIMIT, page));
+        } else {
+            return ok(stream.render(group, posts, postForm, Post.countPostsForGroup(group), LIMIT, page));
+        }
 	}
 	
 	@Transactional(readOnly=true)
 	public static Result media(Long id) {
 		Form<Media> mediaForm = Form.form(Media.class);
 		Group group = Group.findById(id);
-		
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
 		if(!Secured.viewGroup(group)){
 			return redirect(controllers.routes.Application.index());
 		}
-		
-		if (group == null) {
-			return redirect(controllers.routes.GroupController.index());
-		} else {
-			Navigation.set(Level.GROUPS, "Media", group.title, controllers.routes.GroupController.stream(group.id, PAGE));
-			List<Media> mediaSet = group.media; 
-			return ok(media.render(group, mediaForm, mediaSet));
-		}
+
+        Navigation.set(Level.GROUPS, "Media", group.title, controllers.routes.GroupController.stream(group.id, PAGE, false));
+        List<Media> mediaSet = group.media;
+        return ok(media.render(group, mediaForm, mediaSet));
+
 	}
 	
 	public static Result create() {
@@ -141,29 +151,36 @@ public class GroupController extends BaseController {
 			
 			group.createWithGroupAccount(Component.currentAccount());
 			flash("success", successMsg+" erstellt!");
-			return redirect(controllers.routes.GroupController.stream(group.id, PAGE));
+			return redirect(controllers.routes.GroupController.stream(group.id, PAGE, false));
 		}
 	}
 
 	@Transactional
 	public static Result edit(Long id) {
 		Group group = Group.findById(id);
-		
-		
-		if (group == null) {
-			return redirect(controllers.routes.GroupController.index());
-		} else {
-			Navigation.set(Level.GROUPS, "Bearbeiten", group.title, controllers.routes.GroupController.stream(group.id, PAGE));
-			Form<Group> groupForm = Form.form(Group.class).fill(group);
-			groupForm.data().put("type", String.valueOf(group.groupType.ordinal()));
-			return ok(edit.render(group, groupForm));
-		}
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
+
+        Navigation.set(Level.GROUPS, "Bearbeiten", group.title, controllers.routes.GroupController.stream(group.id, PAGE, false));
+        Form<Group> groupForm = Form.form(Group.class).fill(group);
+        groupForm.data().put("type", String.valueOf(group.groupType.ordinal()));
+        return ok(edit.render(group, groupForm));
+
 	}
 	
 	@Transactional
 	public static Result update(Long groupId) {
 		Group group = Group.findById(groupId);
-		Navigation.set(Level.GROUPS, "Bearbeiten", group.title, controllers.routes.GroupController.stream(group.id, PAGE));
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
+
+        Navigation.set(Level.GROUPS, "Bearbeiten", group.title, controllers.routes.GroupController.stream(group.id, PAGE, false));
 		
 		// Check rights
 		if(!Secured.editGroup(group)) {
@@ -200,13 +217,19 @@ public class GroupController extends BaseController {
 		group.description = description;
 		group.update();
 		flash("success", "'" + group.title + "' erfolgreich bearbeitet!");
-		return redirect(controllers.routes.GroupController.stream(groupId, PAGE));
+		return redirect(controllers.routes.GroupController.stream(groupId, PAGE, false));
 		
 	}
 
     @Transactional
 	public static Result delete(Long id) {
 		Group group = Group.findById(id);
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
+
 		if (Secured.deleteGroup(group)) {
 			group.delete();
 			flash("info", "'" + group.title + "' wurde erfolgreich gelöscht!");
@@ -218,19 +241,30 @@ public class GroupController extends BaseController {
 	
 	public static Result token(Long groupId) {
 		Group group = Group.findById(groupId);
-		Navigation.set(Level.GROUPS, "Token eingeben", group.title, controllers.routes.GroupController.stream(group.id, PAGE));
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
+
+		Navigation.set(Level.GROUPS, "Token eingeben", group.title, controllers.routes.GroupController.stream(group.id, PAGE, false));
 		return ok(token.render(group, groupForm));
 	}
 	
 	public static Result validateToken(Long groupId) {
 		Group group = Group.findById(groupId);
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
 		
 		if(Secured.isMemberOfGroup(group, Component.currentAccount())){
 			flash("error", "Du bist bereits Mitglied dieser Gruppe!");
-			return redirect(controllers.routes.GroupController.stream(groupId, PAGE));
+			return redirect(controllers.routes.GroupController.stream(groupId, PAGE, false));
 		}
 		
-		Navigation.set(Level.GROUPS, "Token eingeben", group.title, controllers.routes.GroupController.stream(group.id, PAGE));
+		Navigation.set(Level.GROUPS, "Token eingeben", group.title, controllers.routes.GroupController.stream(group.id, PAGE, false));
 		Form<Group> filledForm = groupForm.bindFromRequest();
 		String enteredToken = filledForm.data().get("token");
 		
@@ -239,7 +273,7 @@ public class GroupController extends BaseController {
 			GroupAccount groupAccount = new GroupAccount(account, group, LinkType.establish);
 			groupAccount.create();
 			flash("success", "Kurs erfolgreich beigetreten!");
-			return redirect(controllers.routes.GroupController.stream(groupId, PAGE));
+			return redirect(controllers.routes.GroupController.stream(groupId, PAGE, false));
 		} else {
 			flash("error", "Hast du dich vielleicht vertippt? Der Token ist leider falsch.");
 			return badRequest(token.render(group, filledForm));
@@ -251,11 +285,16 @@ public class GroupController extends BaseController {
 		Account account = Component.currentAccount();
 		Group group = Group.findById(id);
 		GroupAccount groupAccount;
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
 				
 		if (Secured.isMemberOfGroup(group, account)) {
 			Logger.debug("User is already member of group or course");
 			flash("error", "Du bist bereits Mitglied dieser Gruppe!");
-			return redirect(controllers.routes.GroupController.stream(id, PAGE));
+			return redirect(controllers.routes.GroupController.stream(id, PAGE, false));
 		}
 		
 		// is already requested?
@@ -281,7 +320,7 @@ public class GroupController extends BaseController {
 			groupAccount = new GroupAccount(account, group, LinkType.establish);
 			groupAccount.create();
 			flash("success", "'" + group.title + "' erfolgreich beigetreten!");
-			return redirect(controllers.routes.GroupController.stream(id, PAGE));
+			return redirect(controllers.routes.GroupController.stream(id, PAGE, false));
 		} else if (group.groupType.equals(GroupType.close)) {
 			groupAccount = new GroupAccount(account, group, LinkType.request);
 			groupAccount.create();
@@ -299,6 +338,12 @@ public class GroupController extends BaseController {
 	public static Result removeMember(long groupId, long accountId){
 		Account account = Account.findById(accountId);
 		Group group = Group.findById(groupId);
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
+
 		GroupAccount groupAccount = GroupAccount.find(account, group);
 		
 		Call defaultRedirect = controllers.routes.GroupController.index();
@@ -395,14 +440,26 @@ public class GroupController extends BaseController {
     @Transactional
 	public static Result invite(long groupId) {
 		Group group = Group.findById(groupId);
-		Navigation.set(Level.GROUPS, "Freunde einladen", group.title, controllers.routes.GroupController.stream(group.id, PAGE));
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
+
+		Navigation.set(Level.GROUPS, "Freunde einladen", group.title, controllers.routes.GroupController.stream(group.id, PAGE, false));
 		return ok(invite.render(group, Friendship.friendsToInvite(Component.currentAccount(), group), GroupAccount.findAccountsByGroup(group, LinkType.invite)));
 	}
 
     @Transactional
 	public static Result inviteMember(long groupId) {
 		Group group = Group.findById(groupId);
-		Account currentUser = Component.currentAccount();
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
+
+        Account currentUser = Component.currentAccount();
 		
 		if (Secured.inviteMember(group)) {
             // bind invite list to group
@@ -442,11 +499,17 @@ public class GroupController extends BaseController {
 		}
 		
 		flash("success", Messages.get("group.invite_invited"));
-		return redirect(controllers.routes.GroupController.stream(groupId, PAGE));
+		return redirect(controllers.routes.GroupController.stream(groupId, PAGE, false));
 	}
 	
 	public static Result acceptInvitation(long groupId, long accountId){
 		Group group = Group.findById(groupId);
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
+
 		Account account = Account.findById(accountId);
 		GroupAccount groupAccount = GroupAccount.find(account,group);
 		
@@ -455,11 +518,17 @@ public class GroupController extends BaseController {
 			
 		}
 		
-		return redirect(controllers.routes.GroupController.stream(groupId, PAGE));
+		return redirect(controllers.routes.GroupController.stream(groupId, PAGE, false));
 	}
 	
 	public static Result declineInvitation(long groupId, long accountId){
 		Group group = Group.findById(groupId);
+
+        if (group == null) {
+            flash("error", Messages.get("group.group_not_found"));
+            return redirect(controllers.routes.GroupController.index());
+        }
+        
 		Account account = Account.findById(accountId);
 		GroupAccount groupAccount = GroupAccount.find(account,group);
